requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
pymongo==4.6.1
python-dotenv==1.0.1
huggingface-hub==0.23.2
transformers==4.41.1
torch==2.3.0
Werkzeug==3.0.3
google-generativeai
nltk
requests
